# docs/swagger.yaml
basePath: /api
definitions:
  courseerror.CourseError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  entity.CourseInfo:
    properties:
      cost:
        type: integer
      description:
        type: string
      discount:
        type: integer
      id:
        type: integer
      modules:
        items:
          $ref: '#/definitions/entity.ModuleInfo'
        type: array
      name:
        type: string
      preview:
        type: string
    type: object
  entity.CourseInfoWithPagination:
    properties:
      courses:
        items:
          $ref: '#/definitions/entity.CourseInfo'
        type: array
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
...
  entity.LessonInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: integer
      preview:
        type: string
      video:
        type: string
      watched:
        type: boolean
    type: object
...
  entity.ModuleInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      lessons:
        items:
          $ref: '#/definitions/entity.LessonInfo'
        type: array
      name:
        type: string
      position:
        type: integer
    type: object
  entity.ModuleInfoWithPagination:
    properties:
      modulesInfo:
        items:
          $ref: '#/definitions/entity.ModuleInfo'
        type: array
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
  entity.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pagesCount:
        type: integer
      totalCount:
        type: integer
    type: object
...
host: localhost:70
info:
  contact: {}
  title: Приложение course
  version: "1.0"
paths:
...
  /v1/admin/management/courses:
    get:
      parameters:
      - description: Используется для полного доступа к контенту, при передаче ID
          все остальные параметры игнорируются и происходит проверка на наличие доступа
          к контену
        in: query
        name: id
        type: string
      - description: Название курса
        in: query
        name: name
        type: string
      - description: Описание курса
        in: query
        name: description
        type: string
      - description: Стоимость курса
        in: query
        name: cost
        type: integer
      - description: Размер скидки
        in: query
        name: discount
        type: integer
      - description: Страница
        in: query
        name: page
        required: true
        type: string
      - description: Лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CourseInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Нет доступа к расширенному контенту
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти курсы по фильтрам
      tags:
      - Методы взаимодействия с контентом
...
  /v1/admin/management/lessons:
    get:
      parameters:
      - description: название урока
        in: query
        name: name
        type: string
      - description: описание урока
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: название модуля
        in: query
        name: moduleName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти уроки по фильтрам
      tags:
      - Методы взаимодействия с контентом
  /v1/admin/management/modules:
    get:
      parameters:
      - description: название модуля
        in: query
        name: name
        type: string
      - description: описание модуля
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти модули по фильтрам
      tags:
      - Методы взаимодействия с контентом
...
  /v1/content/courses:
    get:
      parameters:
      - description: Используется для полного доступа к контенту, при передаче ID
          все остальные параметры игнорируются и происходит проверка на наличие доступа
          к контену
        in: query
        name: id
        type: string
      - description: Название курса
        in: query
        name: name
        type: string
      - description: Описание курса
        in: query
        name: description
        type: string
      - description: Стоимость курса
        in: query
        name: cost
        type: integer
      - description: Размер скидки
        in: query
        name: discount
        type: integer
      - description: Страница
        in: query
        name: page
        required: true
        type: string
      - description: Лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CourseInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Нет доступа к расширенному контенту
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти курсы по фильтрам
      tags:
      - Методы взаимодействия с контентом
  /v1/content/lessons:
    get:
      parameters:
      - description: название урока
        in: query
        name: name
        type: string
      - description: описание урока
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: название модуля
        in: query
        name: moduleName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти уроки по фильтрам
      tags:
      - Методы взаимодействия с контентом
  /v1/content/modules:
    get:
      parameters:
      - description: название модуля
        in: query
        name: name
        type: string
      - description: описание модуля
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти модули по фильтрам
      tags:
      - Методы взаимодействия с контентом
...
  /v1/profile/getCourses:
    get:
      parameters:
      - description: Используется для полного доступа к контенту, при передаче ID
          все остальные параметры игнорируются и происходит проверка на наличие доступа
          к контену
        in: query
        name: id
        type: string
      - description: Название курса
        in: query
        name: name
        type: string
      - description: Описание курса
        in: query
        name: description
        type: string
      - description: Стоимость курса
        in: query
        name: cost
        type: integer
      - description: Размер скидки
        in: query
        name: discount
        type: integer
      - description: Страница
        in: query
        name: page
        required: true
        type: string
      - description: Лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CourseInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Нет доступа к расширенному контенту
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти курсы по фильтрам
      tags:
      - Методы взаимодействия с контентом
...
  /v1/profile/lessons:
    get:
      parameters:
      - description: название урока
        in: query
        name: name
        type: string
      - description: описание урока
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: название модуля
        in: query
        name: moduleName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти уроки по фильтрам
      tags:
      - Методы взаимодействия с контентом
  /v1/profile/modules:
    get:
      parameters:
      - description: название модуля
        in: query
        name: name
        type: string
      - description: описание модуля
        in: query
        name: description
        type: string
      - description: название курса
        in: query
        name: courseName
        type: string
      - description: страница
        in: query
        name: page
        required: true
        type: string
      - description: лимит
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ModuleInfoWithPagination'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Найти модули по фильтрам
      tags:
      - Методы взаимодействия с контентом
...
securityDefinitions:
  AdminApiKeyAuth:
    in: cookie
    name: admin_auth
    type: apiKey
  ApiKeyAuth:
    in: cookie
    name: auth
    type: apiKey
swagger: "2.0"